Cast size_t to avoid warnings

+++ src/gfx/font.c
@@ -35,23 +35,23 @@ static int tile_width(TileDescriptor *desc)
 #if SDL_VERSION_ATLEAST(1,3,0)
 	Uint32 key;
 	SDL_GetColorKey(desc->surface->surface, &key);
-#else	
+#else
 	const Uint32 key = desc->surface->surface->format->colorkey;
 #endif
 
 	my_lock(desc->surface->surface);
-	
+
 	int result = 0;
-	
+
 	for (int y = 0 ; y < desc->rect.h ; ++y)
 	{
 		Uint8 *pa = (Uint8 *)desc->surface->surface->pixels + ((int)desc->rect.y + y) * desc->surface->surface->pitch + (int)desc->rect.x * desc->surface->surface->format->BytesPerPixel;
-		
+
 		for (int x = 0 ; x < desc->rect.w ; ++x)
 		{
 			int p = 0;
-			
-			switch (desc->surface->surface->format->BytesPerPixel) 
+
+			switch (desc->surface->surface->format->BytesPerPixel)
 			{
 				default:
 				case 1: p = ((*((Uint8*)pa))) != key; break;
@@ -59,18 +59,18 @@ static int tile_width(TileDescriptor *desc)
 				case 3: p = ((*((Uint32*)pa))&0xffffff) != key; break;
 				case 4: p = ((*((Uint32*)pa))&0xffffff) != key; break;
 			}
-			
+
 			if (p)
 			{
 				result = my_max(result, x);
 			}
-			
+
 			pa += desc->surface->surface->format->BytesPerPixel;
 		}
 	}
-	
+
 	my_unlock(desc->surface->surface);
-	
+
 	return result + 1;
 }
 
@@ -89,9 +89,9 @@ static const TileDescriptor * findchar_slow(const Font *font, char c)
 	const TileDescriptor *tile;
 	for (tile = font->tiledescriptor, tc = font->charmap; *tc ; ++tile, ++tc)
 		if (*tc == c) return tile;
-		
+
 	c = tolower(c);
-		
+
 	for (tile = font->tiledescriptor, tc = font->charmap; *tc ; ++tile, ++tc)
 		if (tolower(*tc) == c) return tile;
 	//debug("Could not find character '%c'", c);
@@ -107,13 +107,13 @@ void font_create(Font *font, GfxSurface *tiles, const int w, const int h, const 
 	font->h = h;
 	font->char_spacing = char_spacing;
 	font->space_width = space_width ? space_width : w;
-	
+
 	if (space_width)
 	{
 		for (int i = 0 ; i < (tiles->surface->w/w)*(tiles->surface->h/h) ; ++i)
 			font->tiledescriptor[i].rect.w = tile_width(&font->tiledescriptor[i]);
 	}
-	
+
 	for (int i = 0 ; i < 256 ; ++i)
 	{
 		font->ordered_tiles[i] = findchar_slow(font, i);
@@ -126,7 +126,7 @@ void font_destroy(Font *font)
 	if (font->tiledescriptor) free(font->tiledescriptor);
 	if (font->charmap) free(font->charmap);
 	if (font->surface) gfx_free_surface(font->surface);
-	
+
 	font->tiledescriptor = NULL;
 	font->charmap = NULL;
 	font->surface = NULL;
@@ -137,14 +137,14 @@ static void inner_write(const Font *font, GfxDomain *dest, const SDL_Rect *r, Ui
 {
 	const char *c = text;
 	int x = (*cursor & 0xff) * font->w, y = ((*cursor >> 8) & 0xff) * font->h, cr = 0, right = dest->screen_w;
-	
+
 	if (r)
 	{
 		x = x + (cr = r->x);
 		y = r->y + y;
 		right = r->w + r->x;
 	}
-		
+
 	for (;*c;++c)
 	{
 		if (*c == '\n' || x >= right)
@@ -153,18 +153,18 @@ static void inner_write(const Font *font, GfxDomain *dest, const SDL_Rect *r, Ui
 			x = cr;
 			*cursor &= (Uint16)0xff00;
 			*cursor += 0x0100;
-			
-			if (*c == '\n') 
+
+			if (*c == '\n')
 				continue;
 		}
-		
+
 		const TileDescriptor *tile = findchar(font, *c);
-		
+
 		if (tile)
 		{
 			SDL_Rect rect = { x, y, tile->rect.w, tile->rect.h };
 			my_BlitSurface(tile->surface, (SDL_Rect*)&tile->rect, dest, &rect);
-			
+
 			if (bounds)
 			{
 				if (bounds->w == 0)
@@ -177,7 +177,7 @@ static void inner_write(const Font *font, GfxDomain *dest, const SDL_Rect *r, Ui
 					bounds->h = my_max(rect.y + rect.h - bounds->y, bounds->h);
 				}
 			}
-			
+
 			x += tile->rect.w + font->char_spacing;
 		}
 		else
@@ -199,10 +199,10 @@ static void inner_write(const Font *font, GfxDomain *dest, const SDL_Rect *r, Ui
 					bounds->h = my_max(y + font->h - bounds->y, bounds->h);
 				}
 			}
-			
+
 			x += font->space_width;
 		}
-		
+
 		++*cursor;
 	}
 }
@@ -218,15 +218,15 @@ void font_write_va(const Font *font, GfxDomain *dest, const SDL_Rect *r, Uint16 
    va_copy( va_cpy, va );
    const int len = vsnprintf(NULL, 0, text, va_cpy) + 1;
    char * formatted = malloc(len * sizeof(*formatted));
-	
+
    va_end( va_cpy );
 #endif
-	
+
 	vsnprintf(formatted, len, text, va);
-	
+
 	inner_write(font, dest, r, cursor, bounds, formatted);
-	
-#ifndef USESTATICTEMPSTRINGS	
+
+#ifndef USESTATICTEMPSTRINGS
 	free(formatted);
 #endif
 }
@@ -235,19 +235,19 @@ void font_write_va(const Font *font, GfxDomain *dest, const SDL_Rect *r, Uint16 
 static int font_load_inner(GfxDomain *domain, Font *font, Bundle *fb)
 {
 	SDL_RWops *rw = SDL_RWFromBundle(fb, "font.bmp");
-	
+
 #ifdef USESDL_IMAGE
 	if (!rw)
 		rw = SDL_RWFromBundle(fb, "font.png");
 #endif
-	
+
 	if (rw)
 	{
 		GfxSurface * s= gfx_load_surface_RW(domain, rw, GFX_KEYED);
-		
+
 		char map[1000];
 		memset(map, 0, sizeof(map));
-		
+
 		{
 			SDL_RWops *rw = SDL_RWFromBundle(fb, "charmap.txt");
 			if (rw)
@@ -256,56 +256,56 @@ static int font_load_inner(GfxDomain *domain, Font *font, Bundle *fb)
 				memset(temp, 0, sizeof(temp));
 				rw->read(rw, temp, 1, sizeof(temp)-1);
 				SDL_RWclose(rw);
-				
+
 				size_t len = my_min(strlen(temp), 256);
 				const char *c = temp;
 				char *m = map;
-				
-				debug("Read charmap (%u chars)", len);
-				
+
+				debug("Read charmap (%lu chars)", (unsigned long)len);
+
 				while (*c)
 				{
 					if (*c == '\\' && len > 1)
 					{
 						char hex = 0;
 						int digits = 0;
-						
+
 						while (len > 1 && digits < 2)
 						{
 							char digit = tolower(*(c + 1));
-							
-							if (!((digit >= '0' && digit <= '9') || (digit >= 'a' && digit <= 'f'))) 
+
+							if (!((digit >= '0' && digit <= '9') || (digit >= 'a' && digit <= 'f')))
 							{
 								if (digits < 1)
 									goto not_hex;
 
 								break;
 							}
-						
+
 							hex <<= 4;
-							
+
 							hex |= (digit >= 'a' ? digit - 'a' + 10 : digit - '0') & 0x0f;
-						
+
 							--len;
 							++digits;
 							++c;
 						}
-						
+
 						if (digits < 1) goto not_hex;
-						
+
 						++c;
-						
+
 						*(m++) = hex;
 					}
 					else
-					{	
+					{
 					not_hex:
 						*(m++) = *(c++);
-							
+
 						--len;
 					}
 				}
-				
+
 				if (*c && len == 0)
 					warning("Excess font charmap chars (max. 256)");
 			}
@@ -315,18 +315,18 @@ static int font_load_inner(GfxDomain *domain, Font *font, Bundle *fb)
 				debug("Charmap not found in font file");
 			}
 		}
-		
+
 		int w, h, spacing = 0, spacewidth = 0;
-		
+
 		{
 			SDL_RWops *rw = SDL_RWFromBundle(fb, "res.txt");
 			char res[10] = { 0 };
-			
+
 			if (rw)
 			{
 				rw->read(rw, res, 1, sizeof(res)-1);
 				SDL_RWclose(rw);
-				
+
 				sscanf(res, "%d %d %d %d", &w, &h, &spacing, &spacewidth);
 			}
 			else
@@ -335,21 +335,21 @@ static int font_load_inner(GfxDomain *domain, Font *font, Bundle *fb)
 				h = 9;
 			}
 		}
-		
+
 		font_create(font, s, w, h, spacing, spacewidth, map);
-	
+
 		font->surface = s;
-		
+
 		bnd_free(fb);
-		
+
 		return 1;
 	}
 	else
 	{
 		warning("Bitmap not found in font file");
-		
+
 		bnd_free(fb);
-		
+
 		return 0;
 	}
 }
@@ -358,13 +358,13 @@ static int font_load_inner(GfxDomain *domain, Font *font, Bundle *fb)
 int font_load_file(GfxDomain *domain, Font *font, char *filename)
 {
 	debug("Loading font '%s'", filename);
-	
-	Bundle fb; 
+
+	Bundle fb;
 	if (bnd_open(&fb, filename))
 	{
 		return font_load_inner(domain, font, &fb);
 	}
-	
+
 	return 0;
 }
 
@@ -374,20 +374,20 @@ int font_load(GfxDomain *domain, Font *font, Bundle *bundle, char *name)
 	debug("Loading font '%s'", name);
 
 	SDL_RWops *rw = SDL_RWFromBundle(bundle, name);
-	
+
 	if (rw)
 	{
 		int r = 0;
-		
+
 		Bundle fb;
-		
+
 		if (bnd_open_RW(&fb, rw))
 		{
 			r = font_load_inner(domain, font, &fb);
 		}
-		
+
 		SDL_RWclose(rw);
-		
+
 		return r;
 	}
 	else
@@ -399,9 +399,9 @@ void font_write_cursor_args(const Font *font, GfxDomain *dest, const SDL_Rect *r
 {
 	va_list va;
 	va_start(va, text);
-	
+
 	font_write_va(font, dest, r, cursor, bounds, text, va);
-	
+
 	va_end(va);
 }
 
@@ -422,12 +422,12 @@ void font_write(const Font *font, GfxDomain *dest, const SDL_Rect *r, const char
 void font_write_args(const Font *font, GfxDomain *dest, const SDL_Rect *r, const char * text, ...)
 {
 	Uint16 cursor = 0;
-	
+
 	va_list va;
 	va_start(va, text);
-	
+
 	font_write_va(font, dest, r, &cursor, NULL, text, va);
-	
+
 	va_end(va);
 }
 
@@ -435,14 +435,14 @@ void font_write_args(const Font *font, GfxDomain *dest, const SDL_Rect *r, const
 int font_load_RW(GfxDomain *domain, Font *font, SDL_RWops *rw)
 {
 	int r = 0;
-		
+
 	Bundle fb;
-	
+
 	if (bnd_open_RW(&fb, rw))
 	{
 		r = font_load_inner(domain, font, &fb);
 	}
-	
+
 	return r;
 }
 
@@ -451,15 +451,15 @@ int font_text_width(const Font *font, const char *text)
 {
 	const char *c = text;
 	int w = 0;
-	
+
 	for (;*c && *c != '\n';++c)
 	{
 		const TileDescriptor *tile = findchar(font, *c);
 		w += tile ? tile->rect.w + font->char_spacing : font->space_width;
 	}
-	
+
 	return w;
-		
+
 }
